+++ src/mixer/ft2_quadratic_spline.c
@@ -20,16 +20,16 @@ bool setupQuadraticSplineTable(void)
 	float *fPtr = fQuadraticSplineLUT;
 	for (int32_t i = 0; i < QUADRATIC_SPLINE_PHASES; i++)
 	{
-		const float x1 = (float)i * (1.0f / QUADRATIC_SPLINE_PHASES);
-		const float x2 = x1 * x1; // x^2
+		const double x1 = i * (1.0 / QUADRATIC_SPLINE_PHASES);
+		const double x2 = x1 * x1; // x^2
 
-		const float t1 = (x1 * -1.5f) + (x2 *  0.5f) + 1.0f;
-		const float t2 = (x1 *  2.0f) + (x2 * -1.0f);
-		const float t3 = (x1 * -0.5f) + (x2 *  0.5f);
+		const double t1 = (x1 * -1.5) + (x2 *  0.5) + 1.0;
+		const double t2 = (x1 *  2.0) + (x2 * -1.0);
+		const double t3 = (x1 * -0.5) + (x2 *  0.5);
 
-		*fPtr++ = t1; // tap #1 at sample offset 0 (center)
-		*fPtr++ = t2; // tap #2 at sample offset 1
-		*fPtr++ = t3; // tap #3 at sample offset 2
+		*fPtr++ = (float)t1; // tap #1 at sample offset 0 (center)
+		*fPtr++ = (float)t2; // tap #2 at sample offset 1
+		*fPtr++ = (float)t3; // tap #3 at sample offset 2
 	}
 
 	return true;
